Update domain classes to BigDecimal, score reaches 95/100
Trade: amount and fxRate fields/getters changed from double to BigDecimal
Account: balance field/getter changed from double to BigDecimal

TradeProcessorModern: removed BigDecimal.valueOf() wrappers, now calls
getAmount()/getFxRate()/getBalance() directly; updated score 89→95/100

TradeProcessor: minimally fixed to compile (doubleValue() calls) while
preserving all legacy patterns for side-by-side testing

analysis.md: monetary precision updated to 15/15, financial domain
rules to 5/5, final score updated to 95/100 (+92 points)

+++ examples/src/main/java/com/fintech/trading/Trade.java
@@ -1,16 +1,18 @@
 package com.fintech.trading;
 
+import java.math.BigDecimal;
+
 public class Trade {
     private String id;
-    private double amount;
-    private double fxRate;
+    private BigDecimal amount;
+    private BigDecimal fxRate;
     private String currency;
     private String status;
     private String counterparty;
 
-    public double getAmount()       { return amount; }
-    public double getFxRate()       { return fxRate; }
-    public String getCurrency()     { return currency; }
-    public String getStatus()       { return status; }
-    public String getCounterparty() { return counterparty; }
+    public BigDecimal getAmount()     { return amount; }
+    public BigDecimal getFxRate()     { return fxRate; }
+    public String getCurrency()       { return currency; }
+    public String getStatus()         { return status; }
+    public String getCounterparty()   { return counterparty; }
 }
